GEODE-3260: Refactoring FunctionCommands
