+++ geode-core/src/test/java/org/apache/geode/management/internal/security/TestCommand.java
@@ -183,8 +183,8 @@ public class TestCommand {
         clusterRead);
     createTestCommand("import cluster-configuration --zip-file-name=value.zip", clusterManage);
 
-    // FunctionCommands
-    // createTestCommand("destroy function --id=InterestCalculations", dataManage);
+    // DestroyFunctionCommand, ExecuteFunctionCommand, ListFunctionCommand
+    createTestCommand("destroy function --id=InterestCalculations", dataManage);
     createTestCommand("execute function --id=InterestCalculations --groups=Group1", dataWrite);
     createTestCommand("list functions", clusterRead);
